Fix longtable's problem with preceding floats
https://www.latex-project.org/cgi-bin/ltxbugs2html.new?pr=tools/3512

"The longtable environment resets \pagegoal to \vsize at the end.  This
behavior is wrong if a document contains a float followed by a longtable
and both the float and the longtable fits on the same page.  In such a
situation, that page would become overfull by exactly the amount taken
up by the float, because LaTeX essentially forgets that it has decided
to place the float on that page."