Re-enable bracketed paste in bash
As part of Bash 5.1, readline introduced a stupid feature called
"faces", which highlights pasted text in a particularly annoying way.
Even worse, this feature is tied to bracketed paste and cannot be
disabled without disabling bracketed paste.
https://bugs.launchpad.net/ubuntu/+source/mate-terminal/+bug/1926256
https://lists.gnu.org/archive/html/info-gnu/2020-12/msg00003.html

Bash 5.1 came to Arch on 20201207, and the last good version of readline
is 8.0.004-1.
https://github.com/archlinux/svntogit-packages/commits/packages/bash/trunk
https://archive.archlinux.org/repos/2020/11/30/core/os/x86_64/
https://archive.archlinux.org/repos/2020/11/30/core/os/x86_64/readline-8.0.004-1-x86_64.pkg.tar.zst